+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/PatternGridCraftingRecipeTransferHandler.java
@@ -1,0 +1,53 @@
+package com.refinedmods.refinedstorage.jei.common;
+
+import com.refinedmods.refinedstorage.common.autocrafting.PatternGridContainerMenu;
+import com.refinedmods.refinedstorage.common.content.Menus;
+import com.refinedmods.refinedstorage.common.support.resource.ItemResource;
+
+import java.util.List;
+import java.util.Optional;
+import javax.annotation.Nullable;
+
+import mezz.jei.api.constants.RecipeTypes;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.recipe.RecipeIngredientRole;
+import mezz.jei.api.recipe.RecipeType;
+import mezz.jei.api.recipe.transfer.IRecipeTransferError;
+import mezz.jei.api.recipe.transfer.IRecipeTransferHandler;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.crafting.CraftingRecipe;
+import net.minecraft.world.item.crafting.RecipeHolder;
+
+class PatternGridCraftingRecipeTransferHandler implements
+    IRecipeTransferHandler<PatternGridContainerMenu, RecipeHolder<CraftingRecipe>> {
+    @Override
+    public Class<? extends PatternGridContainerMenu> getContainerClass() {
+        return PatternGridContainerMenu.class;
+    }
+
+    @Override
+    public Optional<MenuType<PatternGridContainerMenu>> getMenuType() {
+        return Optional.of(Menus.INSTANCE.getPatternGrid());
+    }
+
+    @Override
+    public RecipeType<RecipeHolder<CraftingRecipe>> getRecipeType() {
+        return RecipeTypes.CRAFTING;
+    }
+
+    @Override
+    @Nullable
+    public IRecipeTransferError transferRecipe(final PatternGridContainerMenu containerMenu,
+                                               final RecipeHolder<CraftingRecipe> recipe,
+                                               final IRecipeSlotsView recipeSlots,
+                                               final Player player,
+                                               final boolean maxTransfer,
+                                               final boolean doTransfer) {
+        if (doTransfer) {
+            final List<List<ItemResource>> inputs = SlotUtil.getItems(recipeSlots, RecipeIngredientRole.INPUT);
+            containerMenu.transferCraftingRecipe(inputs);
+        }
+        return null;
+    }
+}