+++ refinedstorage-jei-integration-fabric/src/main/java/com/refinedmods/refinedstorage/jei/fabric/FabricPlatform.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage.jei.fabric;
 
+import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.common.support.resource.FluidResource;
 import com.refinedmods.refinedstorage.jei.common.Platform;
 
@@ -14,6 +15,17 @@ public class FabricPlatform implements Platform {
     public Optional<FluidResource> convertJeiIngredientToFluid(final Object ingredient) {
         if (ingredient instanceof IJeiFluidIngredient fluidIngredient) {
             return Optional.of(ofFluidVariant(fluidIngredient.getFluidVariant()));
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public Optional<ResourceAmount> convertJeiIngredientToFluidAmount(final Object ingredient) {
+        if (ingredient instanceof IJeiFluidIngredient fluidIngredient) {
+            return Optional.of(new ResourceAmount(
+                ofFluidVariant(fluidIngredient.getFluidVariant()),
+                fluidIngredient.getAmount()
+            ));
         }
         return Optional.empty();
     }